auth

+++ backend/controllers/PostController.php
@@ -7,6 +7,7 @@ use common\models\Post;
 use common\models\PostSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\filters\VerbFilter;
 
 /**
@@ -64,6 +65,10 @@ class PostController extends Controller
      */
     public function actionCreate()
     {
+        if (!Yii::$app->user->can('createPost')) {
+            throw new ForbiddenHttpException('对不起，您没有权限访问！');            
+        }
+
         $model = new Post();
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->id]);
@@ -82,6 +87,10 @@ class PostController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (!Yii::$app->user->can('uodatePost')) {
+            throw new ForbiddenHttpException('对不起，您没有权限访问！');            
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -102,6 +111,10 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Yii::$app->user->can('deletePost')) {
+            throw new ForbiddenHttpException('对不起，您没有权限访问！');            
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
